Refactoring entities

+++ src/main/java/com/solarize/solarizeWebBackend/modules/budgetMaterial/BudgetMaterialRepository.java
@@ -2,5 +2,5 @@ package com.solarize.solarizeWebBackend.modules.budgetMaterial;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BudgetMaterialRepository extends JpaRepository<BudgetMaterial, Integer> {
+public interface BudgetMaterialRepository extends JpaRepository<BudgetMaterial, Long> {
 }
